First Version of Analytic Tool
Analytic Packing - Drive & Raw data table plus basic indicators of packed units

+++ node/src/orm/migrations/1720423190876DeletePermissionAddConfidentialityMigration.ts
@@ -2,16 +2,25 @@ import { MigrationInterface, QueryRunner, TableColumn } from "typeorm";
 
 export class DeletePermissionAddConfidentiality1720423190876 implements MigrationInterface {
   public async up(queryRunner: QueryRunner): Promise<void> {
-    await queryRunner.dropColumn("news", "permission");
+    // Check if the "permission" column exists
+    const hasPermissionColumn = await queryRunner.hasColumn("news", "permission");
+    // If "permission" column exists, drop it
+    if (hasPermissionColumn) {
+      await queryRunner.dropColumn("news", "permission");
+    }
 
-    await queryRunner.addColumn(
-      "news",
-      new TableColumn({
-        name: "confidentiality",
-        type: "varchar",
-        default: "'public'",
-      })
-    );
+    // Check if the "confidentiality" column exists before adding it
+    const hasConfidentialityColumn = await queryRunner.hasColumn("news", "confidentiality");
+    if (!hasConfidentialityColumn) {
+      await queryRunner.addColumn(
+        "news",
+        new TableColumn({
+          name: "confidentiality",
+          type: "varchar",
+          default: "'public'",
+        })
+      );
+    }
   }
 
   public async down(queryRunner: QueryRunner): Promise<void> {
